+++ vector/src/main/java/im/vector/riotx/core/extensions/Session.kt
@@ -66,6 +66,6 @@ fun Session.startSyncing(context: Context) {
  * Tell is the session has unsaved e2e keys in the backup
  */
 fun Session.hasUnsavedKeys(): Boolean {
-    return inboundGroupSessionsCount(false) > 0
-            && getKeysBackupService().state != KeysBackupState.ReadyToBackUp
+    return cryptoService().inboundGroupSessionsCount(false) > 0
+            && cryptoService().keysBackupService().state != KeysBackupState.ReadyToBackUp
 }